Refactor page attribute setters
