+++ src/PaginationExtension.ts
@@ -20,7 +20,7 @@ import { isPageNode } from "./utils/nodes/page/page";
 import { getPageNodePosByPageNum } from "./utils/nodes/page/pageNumber";
 import { isValidPaperSize, pageNodeHasPageSize, setPageNodePosPaperSize, setPagePaperSize } from "./utils/paperSize";
 import { getDeviceThemePaperColour, setPageNodePosPaperColour } from "./utils/paperColour";
-import { setBodyNodesAttribute, setPageNodesAttribute } from "./utils/nodes/page/setPageAttributes";
+import { setBodyNodesAttribute, setPageNodesAttribute } from "./utils/nodes/page/attributes/setPageAttributes";
 import { setPageNodePosPaperOrientation } from "./utils/paperOrientation";
 import { isMarginValid, isValidPageMargins, setBodyNodePosPageMargins, updateBodyMargin } from "./utils/pageRegion/margins";
 import { isBorderValid, isValidPageBorders, setPageNodePosPageBorders, updatePageBorder } from "./utils/pageBorders";
